# packages/api/api-swagger-spec.json
{"openapi":"3.0.0","info":{"title":"Wisely Test API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateInventoryDto":{"type":"object","properties":{"startTime":{"format":"date-time","type":"string"},"endTime":{"format":"date-time","type":"string"},"dates":{"type":"array","items":{"type":"string"}},"reservationLimit":{"type":"number"}},"required":["startTime","endTime","dates","reservationLimit"]},"Inventory":{"type":"object","properties":{"id":{"type":"number"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"},"reservationLimit":{"type":"number"},"time":{"format":"date-time","type":"string"},"date":{"format":"date-time","type":"string"},"restaurantId":{"type":"number"}},"required":["id","created","updated","reservationLimit","time","date","restaurantId"]},"Reservation":{"type":"object","properties":{"id":{"type":"number"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"},"name":{"type":"string"},"email":{"type":"string"},"partySize":{"type":"number"},"inventoryId":{"type":"number"},"inventory":{"$ref":"#/components/schemas/Inventory"}},"required":["id","created","updated","name","email","partySize","inventoryId","inventory"]},"CreateReservationDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"partySize":{"type":"number"}},"required":["name","email","partySize"]},"Restaurant":{"type":"object","properties":{"id":{"type":"number"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"},"name":{"type":"string"}},"required":["id","created","updated","name"]},"CreateRestaurantDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]}}},"paths":{"/api/restaurants/{restaurantId}/inventories":{"post":{"operationId":"createInventory","summary":"","parameters":[{"name":"restaurantId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInventoryDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Inventory"}}}}}},"tags":["Inventory"]}},"/api/restaurants/{restaurantId}/reservations":{"get":{"operationId":"getReservations","summary":"Get list of reservations","parameters":[{"name":"restaurantId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}}}},"tags":["Reservations"]},"post":{"operationId":"createReservation","summary":"Create a reservation","parameters":[{"name":"restaurantId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReservationDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}}}},"tags":["Reservations"]}},"/api/restaurants":{"get":{"operationId":"getRestaurants","summary":"Get list of restaurants","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Restaurant"}}}}}},"tags":["Restaurants"]},"post":{"operationId":"createRestaurant","summary":"Create a restaurant","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRestaurantDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Restaurant"}}}}},"tags":["Restaurants"]}},"/api/restaurants/{restaurantId}":{"get":{"operationId":"getRestaurantById","summary":"Get a restaurant by ID","parameters":[{"name":"restaurantId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Restaurant"}}}}},"tags":["Restaurants"]}}}}