# packages/api/src/dtos/create-reservation.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class CreateReservationDto {
    @ApiProperty()
    public name: string;

    @ApiProperty()
    public email: string;

    @ApiProperty()
    public partySize: number;
}